Rename CallNoteSteps view to CallNoteStepsV1
