+++ src/containers/CallNoteStepsContainer.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { createCallNoteWithMentor, changeCallNote } from 'src/actions/callNotes';
 import { getActivity, getCategory, getCall } from 'src/store/helpers';
-import CallNoteSteps from 'src/views/CallNoteSteps';
+import CallNoteStepsV1 from 'src/views/CallNoteStepsV1';
 
 
 export const propsToActions = {
@@ -37,4 +37,4 @@ export const mapStateToProps = (state, { callId }) => {
 };
 
 
-export default connect(mapStateToProps, propsToActions)(CallNoteSteps);
+export default connect(mapStateToProps, propsToActions)(CallNoteStepsV1);
